js/modal.js
Set up a try-catch block to handle errors should they occur during the game. In such an event the open cards will flip back to their closed form. The player can then continue their game without having to restart. This is setup because sometimes mistakes happen that are difficult to forsee or understand why they happen in the first place